Set no_mangle attribute for default exception handlers.
This is needed so that the weak default handlers and
the actual handler by the user can get the same name.
Note that the user handler also needs the no_mangle attribute.
Otherwise the mangled names include crate and module names,
so the names of the two handlers will differ and
the user handler will never be called.